BotTwitter/requirements.txt
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
Pillow==10.3.0
